chore: move datagen classes into proper packages
